move + rename mixin and error if bad table_name given
